+++ src/PHPStan/Laravel/DisallowPartialRouteResource/DisallowPartialRouteFacadeResourceRule.php
@@ -2,11 +2,12 @@
 
 namespace Worksome\CodingStyle\PHPStan\Laravel\DisallowPartialRouteResource;
 
+use Illuminate\Support\Facades\Route;
 use PhpParser\Node;
+use PhpParser\Node\Expr\StaticCall;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\Rule;
 use PHPStan\Rules\RuleError;
-use PHPStan\Rules\RuleErrorBuilder;
 
 /**
  * @implements Rule<Node\Expr\StaticCall>
@@ -28,6 +29,23 @@ class DisallowPartialRouteFacadeResourceRule implements Rule
      */
     public function processNode(Node $node, Scope $scope): array
     {
-        return $this->inspector->inspect($node, $scope);
+        if (! $this->inspector->isApplicable($node)) {
+            return [];
+        }
+
+        if (! $this->isCalledOnRouteFacade($node)) {
+            return [];
+        }
+
+        return $this->inspector->inspect($node);
+    }
+
+    private function isCalledOnRouteFacade(Node $node): bool
+    {
+        if (! $node instanceof StaticCall) {
+            return false;
+        }
+
+        return $node->class->toString() === Route::class;
     }
 }
