+++ src/PHPStan/Laravel/DisallowPartialRouteResource/PartialRouteResourceInspector.php
@@ -2,13 +2,9 @@
 
 namespace Worksome\CodingStyle\PHPStan\Laravel\DisallowPartialRouteResource;
 
-use Illuminate\Support\Facades\Route;
 use PhpParser\Node;
-use PhpParser\Node\Expr\MethodCall;
-use PhpParser\Node\Expr\StaticCall;
 use PHPStan\Rules\RuleError;
 use PHPStan\Rules\RuleErrorBuilder;
-use PHPStan\Analyser\Scope;
 
 final class PartialRouteResourceInspector
 {
@@ -28,12 +24,21 @@ final class PartialRouteResourceInspector
         'only',
     ];
 
-    public function inspect(Node $node, Scope $scope): array
+    public function isApplicable(Node $node): bool
     {
-        if (! $this->isApplicable($node, $scope)) {
-            return [];
+        if (! $node instanceof Node\Expr\CallLike) {
+            return false;
         }
 
+        if (! in_array($node->name->name, $this->resourceMethods)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    public function inspect(Node $node): array
+    {
         $next = $node->getAttribute('next');
 
         while ($next !== null) {
@@ -45,55 +50,6 @@ final class PartialRouteResourceInspector
         }
 
         return [];
-    }
-
-    public function isApplicable(Node $node, Scope $scope): bool
-    {
-        if (! $node instanceof Node\Expr\CallLike) {
-            return false;
-        }
-
-        if (! in_array($node->name->name, $this->resourceMethods)) {
-            return false;
-        }
-
-        if ($this->isCalledOnRouteFacade($node)) {
-            return true;
-        }
-
-        return $this->isCalledInRouteGroupClosure($node, $scope);
-    }
-
-    private function isCalledOnRouteFacade(Node\Expr\CallLike $node): bool
-    {
-        if (! $node instanceof StaticCall) {
-            return false;
-        }
-
-        return $node->class->toString() === Route::class;
-    }
-
-    private function isCalledInRouteGroupClosure(Node\Expr\CallLike $node, Scope $scope): bool
-    {
-        if (! $node instanceof MethodCall) {
-            return false;
-        }
-
-        if (! $scope->isInAnonymousFunction()) {
-            return false;
-        }
-
-        $parent = $node->getAttribute('parent');
-
-        while ($parent !== null) {
-            if ($parent instanceof Node\Expr\StaticCall) {
-                return $parent->class->toString() === Route::class;
-            }
-
-            $parent = $parent->getAttribute('parent');
-        }
-
-        return false;
     }
 
     private function errorFor(string $method): RuleError
